+++ authorization/delivery/http/api.go
@@ -12,6 +12,14 @@ import (
 	"github.com/go-park-mail-ru/2023_2_Vkladyshi/pkg/requests"
 )
 
+type IApi interface {
+	SendResponse(w http.ResponseWriter, response requests.Response)
+	Signin(w http.ResponseWriter, r *http.Request)
+	Signup(w http.ResponseWriter, r *http.Request)
+	LogoutSession(w http.ResponseWriter, r *http.Request)
+	AuthAccept(w http.ResponseWriter, r *http.Request)
+}
+
 type API struct {
 	core usecase.ICore
 	lg   *slog.Logger
